remove referencing in models/aro_hcp
We achieve that by removing any referencing via the
ref statement in the models/aro_hcp definitions. We then
copy all the data types from model/clusters_mgmt, even if not
currently used by ARO-HCP.

However, we remove any locator statements to avoid providing
additional endpoints under the aro_hcp/v1alpha1 API URL.

The copying of all elements will make easier start using them
if needed in the future and also resolving all the attribute
references now that we removed the referencing mechanism.

This will end up adding many data types in the generated Go
models that won't be used and they will be available as data types.
This will also modify existing methods of the generated Go types
to use the duplicated data types. This can have an impact
on the consumers of the models but for aro-hcp case we expect
if they need to change it should not be a big amount of work